Refactor attribute visitors into a base class with derived visitors.
